Telemetry/AVR/Unify serial driver
