+++ radio/src/targets/common/avr/telemetry_driver.cpp
@@ -16,11 +16,22 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
+ *
  */
 
 #include "opentx.h"
 
-#if defined(TELEMETRY_FRSKY)
+#if defined(TELEMETRY_FRSKY) || defined(TELEMETRY_MAVLINK)
+
+enum SERIAL_BAUDS {// KEEP IN SYNC WITH GUI
+  BAUD_4800 = 0,
+  BAUD_9600,
+  BAUD_14400,
+  BAUD_19200,
+  BAUD_38400,
+  BAUD_57600,
+  BAUD_76800
+};
 
 void telemetryEnableTx(void)
 {
@@ -32,6 +43,9 @@ void telemetryEnableRx(void)
   UCSRB_N(TLM_USART) |= (1 << RXEN_N(TLM_USART));  // enable RX
   UCSRB_N(TLM_USART) |= (1 << RXCIE_N(TLM_USART)); // enable Interrupt
 }
+
+void processSerialData(uint8_t data);
+extern uint8_t telemetryRxBufferCount;
 
 ISR(USART_RX_vect_N(TLM_USART))
 {
@@ -79,26 +93,168 @@ ISR(USART_RX_vect_N(TLM_USART))
     telemetryRxBufferCount = 0;
   }
   else {
-    processFrskyTelemetryData(data);
+    processSerialData(data);
   }
 
   cli() ;
   UCSRB_N(TLM_USART) |= (1 << RXCIE_N(TLM_USART)); // enable Interrupt
 }
 
-void telemetryPortInit()
+static void uart_4800(void) {
+  #undef BAUD
+  #define BAUD 4800
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_9600(void) {
+  #undef BAUD
+  #define BAUD 9600
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_14400(void) {
+  #undef BAUD
+  #define BAUD 14400
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_19200(void) {
+  #undef BAUD
+  #define BAUD 19200
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_38400(void) {
+  #undef BAUD
+  #define BAUD 38400
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_57600(void) {
+  #undef BAUD
+  #define BAUD 57600
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+static void uart_76800(void) {
+  #undef BAUD
+  #define BAUD 76800
+  #include <util/setbaud.h>
+  UBRRH_N(TLM_USART) = UBRRH_VALUE;
+  UBRRL_N(TLM_USART) = UBRRL_VALUE;
+#if defined USE_2X
+  UCSRA_N(TLM_USART) |= (1 << U2X_N(TLM_USART)); // enable double speed operation.
+#else
+  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+#endif
+}
+
+void telemetryPortInitFromIndex(uint8_t index) {
+  switch (index) {
+    case BAUD_4800:
+      telemetryPortInit(4800);
+      break;
+    case BAUD_9600:
+      telemetryPortInit(9600);
+      break;
+    case BAUD_14400:
+      telemetryPortInit(14400);
+      break;
+    case BAUD_19200:
+      telemetryPortInit(19200);
+      break;
+    case BAUD_38400:
+      telemetryPortInit(38400);
+      break;
+    case BAUD_57600:
+      telemetryPortInit(57600);
+      break;
+    case BAUD_76800:
+      telemetryPortInit(76800);
+      break;
+  }
+}
+
+void telemetryPortInit() {
+  telemetryPortInit(9600);
+}
+
+void telemetryPortInit(uint32_t baudrate)
 {
 #if !defined(SIMU)
   RXD_DDR_N(TLM_USART) &= ~(1 << RXD_DDR_PIN_N(TLM_USART));   // set RXD pin as input
   RXD_PORT_N(TLM_USART) &= ~(1 << RXD_PORT_PIN_N(TLM_USART)); // disable pullup on RXD pin
-
-  #undef BAUD
-  #define BAUD 9600
-  #include <util/setbaud.h>
-
-  UBRRH_N(TLM_USART) = UBRRH_VALUE;
-  UBRRL_N(TLM_USART) = UBRRL_VALUE;
-  UCSRA_N(TLM_USART) &= ~(1 << U2X_N(TLM_USART)); // disable double speed operation.
+  switch (baudrate) {
+  case 4800:
+    uart_4800();
+    break;
+  case 9600:
+   uart_9600();
+    break;
+  case 14400:
+    uart_14400();
+    break;
+  case 19200:
+    uart_19200();
+    break;
+  case 38400:
+    uart_38400();
+    break;
+  case 57600:
+    uart_57600();
+    break;
+  case 58798:
+    uart_58798();
+    break;
+  case 76800:
+    uart_76800();
+    break;
+  default:
+    uart_57600();
+    break;
+  }
 
   // set 8N1
   UCSRB_N(TLM_USART) = 0 | (0 << RXCIE_N(TLM_USART)) | (0 << TXCIE_N(TLM_USART)) | (0 << UDRIE_N(TLM_USART)) | (0 << RXEN_N(TLM_USART)) | (0 << TXEN_N(TLM_USART)) | (0 << UCSZ2_N(TLM_USART));
@@ -108,9 +264,9 @@ void telemetryPortInit()
   while (UCSRA_N(TLM_USART) & (1 << RXC_N(TLM_USART))) UDR_N(TLM_USART); // flush receive buffer
 
   // These should be running right from power up on a FrSky enabled '9X.
+#if defined(TELEMETRY_FRSKY)
   telemetryEnableTx(); // enable FrSky-Telemetry emission
-  frskyTxBufferCount = 0; // TODO not driver code
-
+#endif
   telemetryEnableRx(); // enable FrSky-Telemetry reception
 #endif
 }
