+++ radio/src/telemetry/frsky.cpp
@@ -48,7 +48,7 @@ enum FrSkyDataState {
 #endif
 };
 
-NOINLINE void processFrskyTelemetryData(uint8_t data)
+NOINLINE void processSerialData(uint8_t data)
 {
   static uint8_t dataState = STATE_DATA_IDLE;
